Use cache

+++ app/src/main/java/io/github/konstantinberkow/pexeltest/cache/DbPhoto.kt
@@ -3,6 +3,6 @@ package io.github.konstantinberkow.pexeltest.cache
 data class DbPhoto(
     val id: Long,
     val authorName: String,
-    val averageColor: Long,
+    val averageColor: Int,
     val src: Map<String, String>,
 )
